+++ src/generate/register.rs
@@ -696,11 +696,6 @@ pub fn fields(
             }
 
             if !derived {
-                let (offset, gen_offset) = if field_dim.is_some() {
-                    (quote! { O }, quote! {, const O: u8 })
-                } else {
-                    (util::unsuffixed(offset as u64), quote! {})
-                };
                 let proxy = if width == 1 {
                     let wproxy = Ident::new(
                         match mwv {
@@ -718,7 +713,7 @@ pub fn fields(
                         },
                         span,
                     );
-                    quote! { crate::#wproxy<'a, #rty, #name_uc_spec, #name_pc_aw, #offset> }
+                    quote! { crate::#wproxy<'a, #rty, #name_uc_spec, #name_pc_aw, O> }
                 } else {
                     let wproxy = Ident::new(
                         if unsafety {
@@ -729,25 +724,17 @@ pub fn fields(
                         span,
                     );
                     let width = &util::unsuffixed(width as _);
-                    quote! { crate::#wproxy<'a, #rty, #name_uc_spec, #fty, #name_pc_aw, #width, #offset> }
+                    quote! { crate::#wproxy<'a, #rty, #name_uc_spec, #fty, #name_pc_aw, #width, O> }
                 };
                 mod_items.extend(quote! {
                     #[doc = #writerdoc]
-                    pub type #name_pc_w<'a #gen_offset> = #proxy;
+                    pub type #name_pc_w<'a, const O: u8> = #proxy;
                 });
             }
             if !proxy_items.is_empty() {
-                mod_items.extend(if field_dim.is_some() {
-                    quote! {
-                        impl<'a, const O: u8> #name_pc_w<'a, O> {
-                            #proxy_items
-                        }
-                    }
-                } else {
-                    quote! {
-                        impl<'a> #name_pc_w<'a> {
-                            #proxy_items
-                        }
+                mod_items.extend(quote! {
+                    impl<'a, const O: u8> #name_pc_w<'a, O> {
+                        #proxy_items
                     }
                 });
             }
@@ -784,10 +771,11 @@ pub fn fields(
                 }
             } else {
                 let doc = description_with_bits(description_raw, offset, width);
+                let offset = util::unsuffixed(offset as u64);
                 w_impl_items.extend(quote! {
                     #[doc = #doc]
                     #inline
-                    pub fn #name_sc(&mut self) -> #name_pc_w {
+                    pub fn #name_sc(&mut self) -> #name_pc_w<#offset> {
                         #name_pc_w::new(self)
                     }
                 });